requirements.txt
PyJWT~=2.10.1
passlib~=1.7.4
fastapi~=0.128.0
cryptography~=46.0.3
SQLAlchemy~=2.0.38
alembic~=1.18.0
pydantic~=2.12.5
uvicorn[standard]~=0.40.0
google-auth~=2.47.0
requests
asyncpg
pydantic[email]
python-multipart
python-dotenv